refactor common popups/toast/listeners UI
