+++ app/src/main/java/com/greenaddress/greenbits/ui/preferences/GaPreferenceActivity.java
@@ -10,8 +10,8 @@ import android.widget.Toast;
 
 import com.greenaddress.greenbits.GaService;
 import com.greenaddress.greenbits.GreenAddressApplication;
+import com.greenaddress.greenbits.ui.UI;
 import com.greenaddress.greenbits.ui.FirstScreenActivity;
-import com.greenaddress.greenbits.ui.GaActivity;
 
 
 // Our GaPreferenceActivity derived classes aren't exported publically, so the
@@ -87,9 +87,9 @@ public abstract class GaPreferenceActivity extends AppCompatPreferenceActivity {
     }
 
     public void toast(final String s) {
-        GaActivity.toastImpl(this, s, Toast.LENGTH_LONG);
+        UI.toast(this, s, Toast.LENGTH_LONG);
     }
     public void toast(final int id) {
-        GaActivity.toastImpl(this, id, Toast.LENGTH_LONG);
+        UI.toast(this, id, Toast.LENGTH_LONG);
     }
 }
